+++ src/enums.ts
@@ -70,7 +70,7 @@ export enum ScoreName {
     FAVOURITED_TAGS = 'FavouritedTags',
     FOLLOWED_ACCOUNTS = 'FollowedAccounts',
     FOLLOWED_TAGS = 'FollowedTags',
-    // FOLLOWERS = 'Followers',
+    FOLLOWERS = 'Followers',
     IMAGE_ATTACHMENTS = 'ImageAttachments',
     INTERACTIONS = 'Interactions',
     MENTIONS_FOLLOWED = 'MentionsFollowed',